+++ src/interfaces/plugins/optionsHandler/simpleOptionsHandler.interface.ts
@@ -1,0 +1,14 @@
+import {OptionsHandlerOptions} from './optionsHandler.interface';
+
+/**
+ * Options for simple options handler plugin
+ */
+export interface SimpleOptionsHandlerOptions extends OptionsHandlerOptions
+{
+    //######################### properties #########################
+
+    /**
+     * Indication whether display selected option in list of available options
+     */
+    listSelected: boolean;
+}