+++ src/plugins/optionsHandler/index.ts
@@ -1,1 +1,1 @@
-export * from './noOptionsHandler/noOptionsHandler.component';
+export * from './simpleOptionsHandler/simpleOptionsHandler.component';